requirements.txt
numpy
opencv-python
tqdm
pytesseract
pandas
PyYAML
